feat: export detecting language to another file
