+++ src/common/index.ts
@@ -2,3 +2,4 @@ export type { default as CalendarEvent } from './CalendarEvent'
 export { default as Swal } from './AppSwal'
 export { default as unimplemented } from './unimplemented'
 export { default as makeCalEvents } from './makeCalEvents'
+export { default as getUserLang } from './userLang'
