+++ src/i18n.ts
@@ -1,18 +1,9 @@
 import { createI18n } from 'vue-i18n'
+import { locale } from '@/common/userLang'
 
 import pl from './locales/pl.json'
 import en from './locales/en.json'
 import ukr from './locales/ukr.json'
-
-const storageLocale = JSON.parse(localStorage.getItem('locale')!)?.locale
-const defaultAppLocale = 'pl'
-
-const possibleLocaleSources = [
-  storageLocale,
-  navigator.language,
-  defaultAppLocale,
-]
-const locale = possibleLocaleSources.find(Boolean)
 
 const i18n = createI18n({
   legacy: false,
